bootstrap: refactor for clarity
