+++ go/provisioner/local/local.go
@@ -20,7 +20,6 @@ import (
 	"errors"
 	"fmt"
 	"log/slog"
-	"net"
 	"os"
 	"os/exec"
 	"path/filepath"
@@ -41,95 +40,6 @@ import (
 	"gopkg.in/yaml.v3"
 )
 
-// CellConfig holds the configuration for a single cell
-type CellConfig struct {
-	Name     string `yaml:"name"`
-	RootPath string `yaml:"root-path"`
-}
-
-// TopologyConfig holds the configuration for cluster topology
-type TopologyConfig struct {
-	Backend        string       `yaml:"backend"`
-	GlobalRootPath string       `yaml:"global-root-path"`
-	Cells          []CellConfig `yaml:"cells"`
-}
-
-// CellServicesConfig holds the service configuration for a specific cell
-type CellServicesConfig struct {
-	Multigateway MultigatewayConfig `yaml:"multigateway"`
-	Multipooler  MultipoolerConfig  `yaml:"multipooler"`
-	Multiorch    MultiorchConfig    `yaml:"multiorch"`
-	Pgctld       PgctldConfig       `yaml:"pgctld"`
-}
-
-// LocalProvisionerConfig represents the typed configuration for the local provisioner
-type LocalProvisionerConfig struct {
-	RootWorkingDir string                        `yaml:"root-working-dir"`
-	DefaultDbName  string                        `yaml:"default-db-name"`
-	Etcd           EtcdConfig                    `yaml:"etcd"`
-	Topology       TopologyConfig                `yaml:"topology"`
-	Multiadmin     MultiadminConfig              `yaml:"multiadmin"`
-	Cells          map[string]CellServicesConfig `yaml:"cells,omitempty"`
-}
-
-// EtcdConfig holds etcd service configuration
-type EtcdConfig struct {
-	Version string `yaml:"version"`
-	DataDir string `yaml:"data-dir"`
-	Port    int    `yaml:"port"`
-}
-
-// MultigatewayConfig holds multigateway service configuration
-type MultigatewayConfig struct {
-	Path     string `yaml:"path"`
-	HttpPort int    `yaml:"http-port"`
-	GrpcPort int    `yaml:"grpc-port"`
-	PgPort   int    `yaml:"pg-port"`
-	LogLevel string `yaml:"log-level"`
-}
-
-// MultipoolerConfig holds multipooler service configuration
-type MultipoolerConfig struct {
-	Path       string `yaml:"path"`
-	Database   string `yaml:"database"`
-	TableGroup string `yaml:"table-group"`
-	ServiceID  string `yaml:"service-id"`
-	PoolerDir  string `yaml:"pooler-dir"` // Directory path for PostgreSQL socket files
-	PgPort     int    `yaml:"pg-port"`    // PostgreSQL port number (same as pgctld)
-	HttpPort   int    `yaml:"http-port"`
-	GrpcPort   int    `yaml:"grpc-port"`
-	LogLevel   string `yaml:"log-level"`
-}
-
-// MultiorchConfig holds multiorch service configuration
-type MultiorchConfig struct {
-	Path     string `yaml:"path"`
-	HttpPort int    `yaml:"http-port"`
-	GrpcPort int    `yaml:"grpc-port"`
-	LogLevel string `yaml:"log-level"`
-}
-
-// MultiadminConfig holds multiadmin service configuration
-type MultiadminConfig struct {
-	Path     string `yaml:"path"`
-	HttpPort int    `yaml:"http-port"`
-	GrpcPort int    `yaml:"grpc-port"`
-	LogLevel string `yaml:"log-level"`
-}
-
-// PgctldConfig holds pgctld service configuration
-type PgctldConfig struct {
-	Path       string `yaml:"path"`
-	PoolerDir  string `yaml:"pooler-dir"`  // Base directory for this pgctld instance
-	GrpcPort   int    `yaml:"grpc-port"`   // gRPC port for pgctld server
-	PgPort     int    `yaml:"pg-port"`     // PostgreSQL port
-	PgDatabase string `yaml:"pg-database"` // PostgreSQL database name
-	PgUser     string `yaml:"pg-user"`     // PostgreSQL username
-	PgPwfile   string `yaml:"pg-pwfile"`   // PostgreSQL password file path (optional)
-	Timeout    int    `yaml:"timeout"`     // Operation timeout in seconds
-	LogLevel   string `yaml:"log-level"`   // Log level
-}
-
 // localProvisioner implements the Provisioner interface for local binary-based provisioning
 type localProvisioner struct {
 	config  *LocalProvisionerConfig
@@ -139,21 +49,6 @@ type localProvisioner struct {
 // Compile-time check to ensure localProvisioner implements Provisioner
 var _ provisioner.Provisioner = (*localProvisioner)(nil)
 
-// LocalProvisionedService represents a service instance that has been provisioned
-type LocalProvisionedService struct {
-	ID         string         `json:"id"`                    // Unique instance ID
-	Service    string         `json:"service"`               // Service name (etcd, multigateway, etc.)
-	PID        int            `json:"pid,omitempty"`         // For binary processes
-	BinaryPath string         `json:"binary-path,omitempty"` // Path to the binary
-	DataDir    string         `json:"data-dir,omitempty"`    // Data directory
-	LogFile    string         `json:"log-file,omitempty"`    // Path to log file
-	Ports      map[string]int `json:"ports"`                 // Port mappings
-	FQDN       string         `json:"fqdn"`                  // Hostname/FQDN
-	Runtime    string         `json:"runtime"`               // "binary"
-	StartedAt  time.Time      `json:"started-at"`            // When it was started
-	Metadata   map[string]any `json:"metadata,omitempty"`    // Additional metadata
-}
-
 const (
 	// StateDir is the directory name where provision state files are stored
 	StateDir = "state"
@@ -162,193 +57,6 @@ const (
 // Name returns the name of this provisioner
 func (p *localProvisioner) Name() string {
 	return "local"
-}
-
-// LoadConfig loads the provisioner-specific configuration from the given config paths
-func (p *localProvisioner) LoadConfig(configPaths []string) error {
-	// Try to find the config file in the provided paths
-	for _, configPath := range configPaths {
-		configFile := filepath.Join(configPath, "multigres.yaml")
-		if _, err := os.Stat(configFile); err == nil {
-			data, err := os.ReadFile(configFile)
-			if err != nil {
-				return fmt.Errorf("failed to read config file %s: %w", configFile, err)
-			}
-
-			// Parse the full config file
-			var fullConfig struct {
-				Provisioner       string         `yaml:"provisioner"`
-				ProvisionerConfig map[string]any `yaml:"provisioner-config,omitempty"`
-			}
-			if err := yaml.Unmarshal(data, &fullConfig); err != nil {
-				return fmt.Errorf("failed to parse config file %s: %w", configFile, err)
-			}
-
-			// Validate that this is for the local provisioner
-			if fullConfig.Provisioner != "local" {
-				return fmt.Errorf("config file %s is for provisioner '%s', not 'local'", configFile, fullConfig.Provisioner)
-			}
-
-			if err := p.ValidateConfig(fullConfig.ProvisionerConfig); err != nil {
-				return fmt.Errorf("failed to validate config file %s: %w", configFile, err)
-			}
-
-			// Convert the provisioner-config section to our typed config
-			yamlData, err := yaml.Marshal(fullConfig.ProvisionerConfig)
-			if err != nil {
-				return fmt.Errorf("failed to marshal provisioner config: %w", err)
-			}
-
-			p.config = &LocalProvisionerConfig{}
-			if err := yaml.Unmarshal(yamlData, p.config); err != nil {
-				return fmt.Errorf("failed to unmarshal provisioner config: %w", err)
-			}
-
-			return nil
-		}
-	}
-
-	return fmt.Errorf("multigres.yaml not found in any of the provided paths: %v", configPaths)
-}
-
-// DefaultConfig returns the default configuration for the local provisioner
-func (p *localProvisioner) DefaultConfig(configPaths []string) map[string]any {
-	baseDir := configPaths[0]
-	binDir, err := getExecutablePath()
-	if err != nil {
-		binDir = "./bin"
-		fmt.Println("Warning: Could not determine executable path, will use ./bin to find binaries")
-	}
-
-	// Generate service IDs for each cell using the same method as topo components
-	serviceIDZone1 := stringutil.RandomString(8)
-	serviceIDZone2 := stringutil.RandomString(8)
-	tableGroup := "default"
-	dbName := "postgres"
-
-	// Create typed configuration with defaults
-	localConfig := LocalProvisionerConfig{
-		RootWorkingDir: baseDir,
-		DefaultDbName:  dbName,
-		Etcd: EtcdConfig{
-			Version: "3.5.9",
-			DataDir: filepath.Join(baseDir, "data", "etcd-data"),
-			Port:    2379,
-		},
-		Topology: TopologyConfig{
-			Backend:        "etcd2",
-			GlobalRootPath: "/multigres/global",
-			Cells: []CellConfig{
-				{
-					Name:     "zone1",
-					RootPath: "/multigres/zone1",
-				},
-				{
-					Name:     "zone2",
-					RootPath: "/multigres/zone2",
-				},
-			},
-		},
-		Multiadmin: MultiadminConfig{
-			Path:     filepath.Join(binDir, "multiadmin"),
-			HttpPort: 15000,
-			GrpcPort: 15990,
-			LogLevel: "info",
-		},
-		Cells: map[string]CellServicesConfig{
-			"zone1": {
-				Multigateway: MultigatewayConfig{
-					Path:     filepath.Join(binDir, "multigateway"),
-					HttpPort: 15001,
-					GrpcPort: 15991,
-					PgPort:   15432,
-					LogLevel: "info",
-				},
-				Multipooler: MultipoolerConfig{
-					Path:       filepath.Join(binDir, "multipooler"),
-					Database:   dbName,
-					TableGroup: tableGroup,
-					ServiceID:  serviceIDZone1,
-					PoolerDir:  GeneratePoolerDir(baseDir, serviceIDZone1),
-					PgPort:     5432, // Same as pgctld for this zone
-					HttpPort:   15100,
-					GrpcPort:   16001,
-					LogLevel:   "info",
-				},
-				Multiorch: MultiorchConfig{
-					Path:     filepath.Join(binDir, "multiorch"),
-					HttpPort: 15300,
-					GrpcPort: 16000,
-					LogLevel: "info",
-				},
-				Pgctld: PgctldConfig{
-					Path:       filepath.Join(binDir, "pgctld"),
-					PoolerDir:  GeneratePoolerDir(baseDir, serviceIDZone1),
-					GrpcPort:   17000,
-					PgPort:     5432,
-					PgDatabase: dbName,
-					PgUser:     "postgres",
-					PgPwfile:   filepath.Join(GeneratePoolerDir(baseDir, serviceIDZone1), "pgpassword.txt"),
-					Timeout:    30,
-					LogLevel:   "info",
-				},
-			},
-			"zone2": {
-				Multigateway: MultigatewayConfig{
-					Path:     filepath.Join(binDir, "multigateway"),
-					HttpPort: 15101, // zone1 + 100
-					GrpcPort: 16091, // zone1 + 100
-					PgPort:   15532, // zone1 + 100
-					LogLevel: "info",
-				},
-				Multipooler: MultipoolerConfig{
-					Path:       filepath.Join(binDir, "multipooler"),
-					Database:   dbName,
-					TableGroup: tableGroup,
-					ServiceID:  serviceIDZone2,
-					PoolerDir:  GeneratePoolerDir(baseDir, serviceIDZone2),
-					PgPort:     5532,  // Same as pgctld for this zone (zone1 + 100)
-					HttpPort:   15200, // zone1 + 100
-					GrpcPort:   16101, // zone1 + 100
-					LogLevel:   "info",
-				},
-				Multiorch: MultiorchConfig{
-					Path:     filepath.Join(binDir, "multiorch"),
-					HttpPort: 15400, // zone1 + 100
-					GrpcPort: 16100, // zone1 + 100
-					LogLevel: "info",
-				},
-				Pgctld: PgctldConfig{
-					Path:       filepath.Join(binDir, "pgctld"),
-					PoolerDir:  GeneratePoolerDir(baseDir, serviceIDZone2),
-					GrpcPort:   17100, // zone1 + 100
-					PgPort:     5532,  // zone1 + 100
-					PgDatabase: dbName,
-					PgUser:     "postgres",
-					PgPwfile:   filepath.Join(GeneratePoolerDir(baseDir, serviceIDZone2), "pgpassword.txt"),
-					Timeout:    30,
-					LogLevel:   "info",
-				},
-			},
-		},
-	}
-
-	// Convert to map[string]any via YAML marshaling to preserve struct ordering
-	yamlData, err := yaml.Marshal(localConfig)
-	if err != nil {
-		// Fallback to empty config if marshaling fails
-		fmt.Printf("Warning: failed to marshal default config: %v\n", err)
-		return map[string]any{}
-	}
-
-	var configMap map[string]any
-	if err := yaml.Unmarshal(yamlData, &configMap); err != nil {
-		// Fallback to empty config if unmarshaling fails
-		fmt.Printf("Warning: failed to unmarshal default config: %v\n", err)
-		return map[string]any{}
-	}
-
-	return configMap
 }
 
 // createPasswordFileAndDirectories creates the pooler directory structure and password file
@@ -638,26 +346,6 @@ func (p *localProvisioner) readServiceLogs(logFile string, lines int) string {
 	return result
 }
 
-// validateProcessRunning checks if a process with the given PID is still running
-func (p *localProvisioner) validateProcessRunning(pid int) error {
-	if pid <= 0 {
-		return fmt.Errorf("invalid PID: %d", pid)
-	}
-
-	process, err := os.FindProcess(pid)
-	if err != nil {
-		return fmt.Errorf("process with PID %d not found: %w", pid, err)
-	}
-
-	// Send signal 0 to check if process exists without actually sending a signal
-	err = process.Signal(syscall.Signal(0))
-	if err != nil {
-		return fmt.Errorf("process with PID %d is not running: %w", pid, err)
-	}
-
-	return nil
-}
-
 // getRootWorkingDir returns the root working directory from config
 func (p *localProvisioner) getRootWorkingDir() string {
 	if p.config == nil {
@@ -670,119 +358,6 @@ func (p *localProvisioner) getRootWorkingDir() string {
 // GeneratePoolerDir generates a pooler directory path for a given base directory and service ID
 func GeneratePoolerDir(baseDir, serviceID string) string {
 	return filepath.Join(baseDir, "data", fmt.Sprintf("pooler_%s", serviceID))
-}
-
-// getStateDir returns the path to the state directory
-func (p *localProvisioner) getStateDir() string {
-	return filepath.Join(p.getRootWorkingDir(), "state")
-}
-
-// getLogsDir returns the path to the logs directory
-func (p *localProvisioner) getLogsDir() string {
-	return filepath.Join(p.getRootWorkingDir(), "logs")
-}
-
-func (p *localProvisioner) getDataDir() string {
-	return filepath.Join(p.getRootWorkingDir(), "data")
-}
-
-// createLogFile creates a log file path and ensures the directory exists
-func (p *localProvisioner) createLogFile(serviceName, serviceID, databaseName string) (string, error) {
-	logsDir := p.getLogsDir()
-	var serviceLogDir string
-
-	if databaseName != "" {
-		// For database services: logs/dbs/dbname/servicename
-		serviceLogDir = filepath.Join(logsDir, "dbs", databaseName, serviceName)
-	} else {
-		// For non-database services (like etcd): logs/servicename
-		serviceLogDir = filepath.Join(logsDir, serviceName)
-	}
-
-	// Create the service-specific log directory
-	if err := os.MkdirAll(serviceLogDir, 0o755); err != nil {
-		return "", fmt.Errorf("failed to create log directory %s: %w", serviceLogDir, err)
-	}
-
-	// Create the log file path
-	logFile := filepath.Join(serviceLogDir, fmt.Sprintf("%s.log", serviceID))
-	return logFile, nil
-}
-
-// cleanupLogFile removes a log file if it exists
-func (p *localProvisioner) cleanupLogFile(logFilePath string) error {
-	if logFilePath == "" {
-		return nil
-	}
-
-	// Check if log file exists
-	if _, err := os.Stat(logFilePath); os.IsNotExist(err) {
-		return nil // File doesn't exist, nothing to clean up
-	}
-
-	// Remove the log file
-	if err := os.Remove(logFilePath); err != nil {
-		return fmt.Errorf("failed to remove log file %s: %w", logFilePath, err)
-	}
-
-	fmt.Printf("Cleaned up log file: %s\n", logFilePath)
-	return nil
-}
-
-// saveServiceState saves the provisioned service state to disk
-func (p *localProvisioner) saveServiceState(service *LocalProvisionedService, databaseName string) error {
-	stateDir := p.getStateDir()
-	var targetDir string
-
-	if databaseName != "" {
-		// For database services: state/dbs/dbname
-		targetDir = filepath.Join(stateDir, "dbs", databaseName)
-	} else {
-		// For non-database services (like etcd): state/
-		targetDir = stateDir
-	}
-
-	if err := os.MkdirAll(targetDir, 0o755); err != nil {
-		return fmt.Errorf("failed to create state directory %s: %w", targetDir, err)
-	}
-
-	// File name format: service_id.json (e.g., etcd_abc123.json)
-	fileName := fmt.Sprintf("%s_%s.json", service.Service, service.ID)
-	filePath := filepath.Join(targetDir, fileName)
-
-	data, err := json.MarshalIndent(service, "", "  ")
-	if err != nil {
-		return fmt.Errorf("failed to marshal service state: %w", err)
-	}
-
-	if err := os.WriteFile(filePath, data, 0o644); err != nil {
-		return fmt.Errorf("failed to write service state file %s: %w", filePath, err)
-	}
-
-	return nil
-}
-
-// removeServiceState removes a service state file from disk
-func (p *localProvisioner) removeServiceState(serviceID, serviceName, databaseName string) error {
-	stateDir := p.getStateDir()
-	var targetDir string
-
-	if databaseName != "" {
-		// For database services: state/dbs/dbname
-		targetDir = filepath.Join(stateDir, "dbs", databaseName)
-	} else {
-		// For non-database services (like etcd): state/
-		targetDir = stateDir
-	}
-
-	fileName := fmt.Sprintf("%s_%s.json", serviceName, serviceID)
-	filePath := filepath.Join(targetDir, fileName)
-
-	if err := os.Remove(filePath); err != nil && !os.IsNotExist(err) {
-		return fmt.Errorf("failed to remove state file %s: %w", filePath, err)
-	}
-
-	return nil
 }
 
 // provisionMultigateway provisions multigateway using either binaries or Docker containers
@@ -1443,349 +1018,6 @@ func (p *localProvisioner) loadServiceState(req *provisioner.DeprovisionRequest)
 	return &service, nil
 }
 
-// loadDbProvisionedServices loads provisioned services for a specific database
-func (p *localProvisioner) loadDbProvisionedServices(databaseName string) ([]*LocalProvisionedService, error) {
-	if databaseName == "" {
-		return nil, fmt.Errorf("database name is required")
-	}
-
-	stateDir := p.getStateDir()
-	targetDir := filepath.Join(stateDir, "dbs", databaseName)
-
-	// Check if target directory exists
-	if _, err := os.Stat(targetDir); os.IsNotExist(err) {
-		return nil, nil // No services for this database
-	}
-
-	entries, err := os.ReadDir(targetDir)
-	if err != nil {
-		return nil, fmt.Errorf("failed to read directory %s: %w", targetDir, err)
-	}
-
-	var services []*LocalProvisionedService
-	for _, entry := range entries {
-		if !entry.IsDir() && strings.HasSuffix(entry.Name(), ".json") {
-			// Parse filename: service_id.json
-			name := strings.TrimSuffix(entry.Name(), ".json")
-			parts := strings.SplitN(name, "_", 2)
-			if len(parts) != 2 {
-				// Skip invalid state files
-				continue
-			}
-
-			serviceName := parts[0]
-			serviceID := parts[1]
-
-			req := &provisioner.DeprovisionRequest{
-				Service:      serviceName,
-				ServiceID:    serviceID,
-				DatabaseName: databaseName,
-			}
-			service, err := p.loadServiceState(req)
-			if err != nil {
-				// Log warning but continue with other services
-				fmt.Printf("Warning: failed to load state for %s service %s: %v\n", serviceName, serviceID, err)
-				continue
-			}
-
-			if service != nil {
-				services = append(services, service)
-			}
-		}
-	}
-
-	return services, nil
-}
-
-// loadEtcdServices loads etcd services from the top-level state directory
-func (p *localProvisioner) loadEtcdServices() ([]*LocalProvisionedService, error) {
-	stateDir := p.getStateDir()
-
-	// Check if state directory exists
-	if _, err := os.Stat(stateDir); os.IsNotExist(err) {
-		return nil, nil // No state directory, no services running
-	}
-
-	entries, err := os.ReadDir(stateDir)
-	if err != nil {
-		return nil, fmt.Errorf("failed to read directory %s: %w", stateDir, err)
-	}
-
-	var services []*LocalProvisionedService
-	for _, entry := range entries {
-		if !entry.IsDir() && strings.HasSuffix(entry.Name(), ".json") && strings.HasPrefix(entry.Name(), "etcd_") {
-			// Parse filename: etcd_serviceID.json
-			name := strings.TrimSuffix(entry.Name(), ".json")
-			parts := strings.SplitN(name, "_", 2)
-			if len(parts) != 2 || parts[0] != "etcd" {
-				continue
-			}
-
-			serviceID := parts[1]
-
-			req := &provisioner.DeprovisionRequest{
-				Service:      "etcd",
-				ServiceID:    serviceID,
-				DatabaseName: "", // etcd is a global service
-			}
-			service, err := p.loadServiceState(req)
-			if err != nil {
-				// Log warning but continue with other services
-				fmt.Printf("Warning: failed to load state for etcd service %s: %v\n", serviceID, err)
-				continue
-			}
-
-			if service != nil {
-				services = append(services, service)
-			}
-		}
-	}
-
-	return services, nil
-}
-
-// loadGlobalServices loads all global services (non-database services) from state files
-func (p *localProvisioner) loadGlobalServices() ([]*LocalProvisionedService, error) {
-	stateDir := p.getStateDir()
-
-	// Check if state directory exists
-	if _, err := os.Stat(stateDir); os.IsNotExist(err) {
-		return nil, nil // No state directory, no services running
-	}
-
-	entries, err := os.ReadDir(stateDir)
-	if err != nil {
-		return nil, fmt.Errorf("failed to read directory %s: %w", stateDir, err)
-	}
-
-	var services []*LocalProvisionedService
-	for _, entry := range entries {
-		if !entry.IsDir() && strings.HasSuffix(entry.Name(), ".json") {
-			// Parse filename: serviceName_serviceID.json
-			name := strings.TrimSuffix(entry.Name(), ".json")
-			parts := strings.SplitN(name, "_", 2)
-			if len(parts) != 2 {
-				continue
-			}
-
-			serviceName := parts[0]
-			serviceID := parts[1]
-
-			// Load global services (non-etcd services can be included here)
-			if serviceName == "multiadmin" || serviceName == "etcd" {
-				req := &provisioner.DeprovisionRequest{
-					Service:      serviceName,
-					ServiceID:    serviceID,
-					DatabaseName: "", // global services have no database name
-				}
-
-				service, err := p.loadServiceState(req)
-				if err != nil {
-					// Log warning but continue with other services
-					fmt.Printf("Warning: failed to load state for global service %s: %v\n", serviceID, err)
-					continue
-				}
-
-				if service != nil {
-					services = append(services, service)
-				}
-			}
-		}
-	}
-
-	return services, nil
-}
-
-// findRunningDbService finds a running service by service name within a specific database and cell
-func (p *localProvisioner) findRunningDbService(serviceName, databaseName, cell string) (*LocalProvisionedService, error) {
-	services, err := p.loadDbProvisionedServices(databaseName)
-	if err != nil {
-		return nil, fmt.Errorf("failed to load service states for database %s: %w", databaseName, err)
-	}
-
-	for _, service := range services {
-		if service.Service == serviceName {
-			// Check if the service matches the cell
-			if serviceCell, ok := service.Metadata["cell"].(string); ok && serviceCell == cell {
-				// Check if the service is actually still running
-				if service.PID > 0 {
-					if err := p.validateProcessRunning(service.PID); err == nil {
-						return service, nil
-					}
-				}
-			}
-		}
-	}
-
-	// Check for port conflicts with other processes using cell-specific config
-	expectedPorts := p.getExpectedPortsForDbService(serviceName, cell)
-	for portName, port := range expectedPorts {
-		if err := p.checkPortConflict(port, serviceName, portName); err != nil {
-			return nil, err
-		}
-	}
-
-	return nil, nil // No running service found
-}
-
-// getExpectedPortsForDbService returns expected ports for a DB-scoped service (per cell)
-func (p *localProvisioner) getExpectedPortsForDbService(serviceName, cell string) map[string]int {
-	ports := make(map[string]int)
-
-	cellConfig, err := p.getCellServiceConfig(cell, serviceName)
-	if err != nil {
-		return ports
-	}
-
-	switch serviceName {
-	case "multigateway":
-		if httpPort, ok := cellConfig["http_port"].(int); ok {
-			ports["http"] = httpPort
-		}
-		if grpcPort, ok := cellConfig["grpc_port"].(int); ok {
-			ports["grpc"] = grpcPort
-		}
-	case "multipooler":
-		if grpcPort, ok := cellConfig["grpc_port"].(int); ok {
-			ports["grpc"] = grpcPort
-		}
-		if httpPort, ok := cellConfig["http_port"].(int); ok && httpPort > 0 {
-			ports["http"] = httpPort
-		}
-	case "multiorch":
-		if grpcPort, ok := cellConfig["grpc_port"].(int); ok {
-			ports["grpc"] = grpcPort
-		}
-		if httpPort, ok := cellConfig["http_port"].(int); ok && httpPort > 0 {
-			ports["http"] = httpPort
-		}
-	}
-
-	return ports
-}
-
-// getExpectedPortsForService returns the expected ports for a service based on its configuration
-func (p *localProvisioner) getExpectedPortsForService(serviceName string) map[string]int {
-	serviceConfig := p.getServiceConfig(serviceName)
-	ports := make(map[string]int)
-
-	switch serviceName {
-	case "multigateway":
-		if httpPort, ok := serviceConfig["http_port"].(int); ok {
-			ports["http"] = httpPort
-		}
-		if grpcPort, ok := serviceConfig["grpc_port"].(int); ok {
-			ports["grpc"] = grpcPort
-		}
-	case "multipooler":
-		if grpcPort, ok := serviceConfig["grpc_port"].(int); ok {
-			ports["grpc"] = grpcPort
-		}
-		if httpPort, ok := serviceConfig["http_port"].(int); ok && httpPort > 0 {
-			ports["http"] = httpPort
-		}
-	case "multiorch":
-		if grpcPort, ok := serviceConfig["grpc_port"].(int); ok {
-			ports["grpc"] = grpcPort
-		}
-		if httpPort, ok := serviceConfig["http_port"].(int); ok && httpPort > 0 {
-			ports["http"] = httpPort
-		}
-	case "multiadmin":
-		if httpPort, ok := serviceConfig["http_port"].(int); ok {
-			ports["http"] = httpPort
-		}
-		if grpcPort, ok := serviceConfig["grpc_port"].(int); ok {
-			ports["grpc"] = grpcPort
-		}
-	case "etcd":
-		if port, ok := serviceConfig["port"].(int); ok {
-			ports["tcp"] = port
-		}
-	}
-
-	return ports
-}
-
-// checkPortConflict checks if a port is already in use by another process
-func (p *localProvisioner) checkPortConflict(port int, serviceName, portName string) error {
-	if port <= 0 {
-		return nil // Skip invalid ports
-	}
-
-	address := fmt.Sprintf("localhost:%d", port)
-	conn, err := net.DialTimeout("tcp", address, 1*time.Second)
-	if err != nil {
-		// Port is not in use, this is good
-		return nil
-	}
-	conn.Close()
-
-	// Port is in use by some process
-	return fmt.Errorf("cannot start %s: port %d (%s) is already in use by another process. "+
-		"There is no way to do a clean start. Please kill the process using port %d or change the configuration",
-		serviceName, port, portName, port)
-}
-
-// findRunningEtcdService finds a running etcd service
-func (p *localProvisioner) findRunningEtcdService() (*LocalProvisionedService, error) {
-	services, err := p.loadEtcdServices()
-	if err != nil {
-		return nil, fmt.Errorf("failed to load etcd service states: %w", err)
-	}
-
-	for _, service := range services {
-		if service.Service == "etcd" {
-			// Check if the service is actually still running
-			if service.PID > 0 {
-				if err := p.validateProcessRunning(service.PID); err == nil {
-					return service, nil
-				}
-			}
-		}
-	}
-
-	// Check for port conflicts with other processes
-	expectedPorts := p.getExpectedPortsForService("etcd")
-	for portName, port := range expectedPorts {
-		if err := p.checkPortConflict(port, "etcd", portName); err != nil {
-			return nil, err
-		}
-	}
-
-	return nil, nil // No running etcd service found
-}
-
-// findRunningService finds a running service by service name (for global services like multiadmin)
-func (p *localProvisioner) findRunningService(serviceName string) (*LocalProvisionedService, error) {
-	// Load global services (e.g., multiadmin, etcd)
-	services, err := p.loadGlobalServices()
-	if err != nil {
-		return nil, fmt.Errorf("failed to load global service states: %w", err)
-	}
-
-	for _, service := range services {
-		if service.Service == serviceName {
-			// Check if the service is actually still running
-			if service.PID > 0 {
-				if err := p.validateProcessRunning(service.PID); err == nil {
-					return service, nil
-				}
-			}
-		}
-	}
-
-	// Check for port conflicts with other processes
-	expectedPorts := p.getExpectedPortsForService(serviceName)
-	for portName, port := range expectedPorts {
-		if err := p.checkPortConflict(port, serviceName, portName); err != nil {
-			return nil, err
-		}
-	}
-
-	return nil, nil // No running service found
-}
-
 // stopService stops a specific service based on its type using the internal methods
 func (p *localProvisioner) stopService(ctx context.Context, req *provisioner.DeprovisionRequest) error {
 	switch req.Service {
@@ -2392,33 +1624,6 @@ func (p *localProvisioner) getCellIndex(cellName string) (int, error) {
 	return -1, fmt.Errorf("cell %s not found", cellName)
 }
 
-// calculatePortOffset calculates the port offset for a given cell (100 per cell)
-func (p *localProvisioner) calculatePortOffset(cellName string) (int, error) {
-	cellIndex, err := p.getCellIndex(cellName)
-	if err != nil {
-		return 0, err
-	}
-	return cellIndex * 100, nil
-}
-
-// getDefaultCell returns the first cell configuration (for backward compatibility)
-func (p *localProvisioner) getDefaultCell() (string, *CellConfig, error) {
-	if p.config == nil {
-		return "", nil, fmt.Errorf("provisioner config not set")
-	}
-
-	if len(p.config.Topology.Cells) == 0 {
-		return "", nil, fmt.Errorf("no cells configured")
-	}
-
-	// Return the first cell (for backward compatibility with single-cell setup)
-	for _, cell := range p.config.Topology.Cells {
-		return cell.Name, &cell, nil
-	}
-
-	return "", nil, fmt.Errorf("no cells found")
-}
-
 // getCellByName returns the cell configuration for a specific cell name
 func (p *localProvisioner) getCellByName(cellName string) (*CellConfig, error) {
 	if p.config == nil {
@@ -2478,80 +1683,6 @@ func (p *localProvisioner) ValidateConfig(config map[string]any) error {
 	return nil
 }
 
-// getServiceConfig gets the configuration for a specific service (global services only)
-func (p *localProvisioner) getServiceConfig(service string) map[string]any {
-	switch service {
-	case "etcd":
-		return map[string]any{
-			"version":  p.config.Etcd.Version,
-			"data-dir": p.config.Etcd.DataDir,
-			"port":     p.config.Etcd.Port,
-		}
-	case "multiadmin":
-		return map[string]any{
-			"path":      p.config.Multiadmin.Path,
-			"http_port": p.config.Multiadmin.HttpPort,
-			"grpc_port": p.config.Multiadmin.GrpcPort,
-			"log_level": p.config.Multiadmin.LogLevel,
-		}
-	default:
-		// Return empty config if not found
-		return map[string]any{}
-	}
-}
-
-// getCellServiceConfig gets the configuration for a specific service in a specific cell
-func (p *localProvisioner) getCellServiceConfig(cellName, service string) (map[string]any, error) {
-	cellServices, exists := p.config.Cells[cellName]
-	if !exists {
-		return nil, fmt.Errorf("cell %s not found in configuration", cellName)
-	}
-
-	switch service {
-	case "multigateway":
-		return map[string]any{
-			"path":      cellServices.Multigateway.Path,
-			"http_port": cellServices.Multigateway.HttpPort,
-			"grpc_port": cellServices.Multigateway.GrpcPort,
-			"pg_port":   cellServices.Multigateway.PgPort,
-			"log_level": cellServices.Multigateway.LogLevel,
-		}, nil
-	case "multipooler":
-		return map[string]any{
-			"path":        cellServices.Multipooler.Path,
-			"database":    cellServices.Multipooler.Database,
-			"table_group": cellServices.Multipooler.TableGroup,
-			"service-id":  cellServices.Multipooler.ServiceID,
-			"http_port":   cellServices.Multipooler.HttpPort,
-			"grpc_port":   cellServices.Multipooler.GrpcPort,
-			"log_level":   cellServices.Multipooler.LogLevel,
-			"pooler_dir":  cellServices.Multipooler.PoolerDir,
-			"pg_port":     cellServices.Multipooler.PgPort,
-		}, nil
-	case "multiorch":
-		return map[string]any{
-			"path":      cellServices.Multiorch.Path,
-			"http_port": cellServices.Multiorch.HttpPort,
-			"grpc_port": cellServices.Multiorch.GrpcPort,
-			"log_level": cellServices.Multiorch.LogLevel,
-		}, nil
-	case "pgctld":
-		return map[string]any{
-			"path":        cellServices.Pgctld.Path,
-			"pooler_dir":  cellServices.Pgctld.PoolerDir,
-			"grpc_port":   cellServices.Pgctld.GrpcPort,
-			"pg_port":     cellServices.Pgctld.PgPort,
-			"pg_database": cellServices.Pgctld.PgDatabase,
-			"pg_user":     cellServices.Pgctld.PgUser,
-			"pg_pwfile":   cellServices.Pgctld.PgPwfile,
-			"timeout":     cellServices.Pgctld.Timeout,
-			"log_level":   cellServices.Pgctld.LogLevel,
-		}, nil
-	default:
-		return nil, fmt.Errorf("unknown service %s", service)
-	}
-}
-
 // NewLocalProvisioner creates a new local provisioner instance
 func NewLocalProvisioner() (provisioner.Provisioner, error) {
 	p := &localProvisioner{
@@ -2559,15 +1690,6 @@ func NewLocalProvisioner() (provisioner.Provisioner, error) {
 	}
 
 	return p, nil
-}
-
-// getMapKeys returns the keys of a map for debugging purposes
-func getMapKeys(m map[string]interface{}) []string {
-	keys := make([]string, 0, len(m))
-	for k := range m {
-		keys = append(keys, k)
-	}
-	return keys
 }
 
 func getExecutablePath() (string, error) {
